좋아요 개수 반환 이해하기 쉽게 변경

+++ src/main/java/com/springstudy/backend/API/Repository/ReviewLikeRepository.java
@@ -13,6 +13,6 @@ public interface ReviewLikeRepository extends JpaRepository<ReviewLike, ReviewLi
     void deleteByUserIdAndReviewId(Long userId, Long reviewId);
 
     @Query("SELECT COUNT(rl) FROM ReviewLike rl WHERE rl.review.id = :reviewId")
-    long countByReviewId(@Param("reviewId") Long reviewId);
+    int countByReviewId(@Param("reviewId") Long reviewId);
 
 }
